fix(micrometer): :zap: use single-precision float throughout application
All double-precision arithmetic was software-emulated, wasting cycles,
but patching workaround "double-precision math".

Convert all sensor data, state machine inputs, Kalman filter internals,
and baro altitude computation from double to float. Adjust math calls
(pow->powf, sqrt->sqrtf, atan2->atan2f, fabs->fabsf), suffix literals,
and widen test tolerances to account for reduced powf precision.
Kalman filter epsilon changed from 1e-12 to 1e-6f (float min ~1e-7).

This change is possible since CONFIG_FPU and CONFIG_FPU_SHARED are set
for micrometer.

Workaround: double-precision math - fixed

+++ include/aurora/lib/baro.h
@@ -56,7 +56,7 @@ int baro_init(const struct device *dev);
  * @retval 0 on success.
  * @retval -EINVAL if @p ref_kpa is not positive.
  */
-int baro_set_reference(double ref_kpa);
+int baro_set_reference(float ref_kpa);
 
 /**
  * @brief Convert a pressure reading to altitude AGL.
@@ -68,7 +68,7 @@ int baro_set_reference(double ref_kpa);
  *
  * @return Altitude in meters above the reference level.
  */
-double baro_pressure_to_altitude(double press_kpa);
+float baro_pressure_to_altitude(float press_kpa);
 
 /** @} */
